EOS-22001: Bug fixes for PerfBot

+++ performance/PerfBot/src/data_parser.py
@@ -8,6 +8,7 @@ e.g. python 3 ./parser.py './Data/hsbench/hsbench.log' './Data/cosbench/s3-5050r
 # imports
 import os
 import time
+import yaml
 
 from Parsers.hsbench_parser import extract_HSBench_logs, convert_HSlogs_to_JSON
 from Parsers.cosbench_parser import convert_COSlogs_to_JSON
@@ -18,7 +19,10 @@ from Parsers.logs_parser import parse_errors
 # variables
 input_folder_path = './Input/'
 parsed_data_path = './ParsedData/'
-quantum = 5
+
+with open("perfbot_config.yml", "r") as perfbot_config:
+    configs = yaml.safe_load(perfbot_config)
+    quantum = configs['quantum']
 
 
 def check_os_paths():
@@ -44,7 +48,8 @@ def run_cosbench_parser(run_ID, COS_source_file_path, cos_obj_size):
     # COS Bench parser
     COS_input_file_path = './Input/cosbench.json'
 
-    convert_COSlogs_to_JSON(run_ID, COS_source_file_path, COS_input_file_path, cos_obj_size)
+    convert_COSlogs_to_JSON(run_ID, COS_source_file_path,
+                            COS_input_file_path, cos_obj_size)
     time.sleep(1)
 
 
@@ -62,17 +67,32 @@ def run_s3bench_parser(run_ID, S3_source_file_path):
 def parse_data(run_ID, run_dirs, log_files, cos_obj_size):
     check_os_paths()
 
-    run_hsbench_parser(run_ID, run_dirs[0])
-    print("~ collected performance data from HSBench run directories")
+    if run_dirs[0] is not None:
+        run_hsbench_parser(run_ID, run_dirs[0])
+        print("~ collected performance data from HSBench run directories")
+    else:
+        print("~ HSbench run data file is not given, skipping...")
 
-    run_cosbench_parser(run_ID, run_dirs[1], cos_obj_size)
-    print("~ collected performance data from COSBench run directories")
+    if run_dirs[1] is not None:
+        if cos_obj_size is None:
+            print("~ object size is not provided for COSbench, skipping...")
+        else:
+            run_cosbench_parser(run_ID, run_dirs[1], cos_obj_size)
+            print("~ collected performance data from COSBench run directories")
+    else:
+        print("~ COSbench run data file is not given, skipping...")
 
-    run_s3bench_parser(run_ID, run_dirs[2])
-    print("~ collected performance data from S3Bench run directories")
+    if run_dirs[2] is not None:
+        run_s3bench_parser(run_ID, run_dirs[2])
+        print("~ collected performance data from S3Bench run directories")
+    else:
+        print("~ S3bench run data file is not given, skipping...")
 
-    print("~ Parsing run logs...")
-    parse_errors(run_ID, log_files[0], log_files[1], log_files[2])
+    if log_files[0] is not None and log_files[1] is not None and log_files[2] is not None:
+        print("~ Parsing run logs...")
+        parse_errors(run_ID, log_files[0], log_files[1], log_files[2])
 
-    print("~ Collected Error logs and pushed to database")
-    os.removedirs(parsed_data_path)
+        print("\n~ Collected Error logs and pushed them to the database")
+        os.removedirs(parsed_data_path)
+    else:
+        print("~ Run log files are not provided for at least 1 tool, skipping...")
